Organizzazione progetto e preparazione per il merge con master
