+++ src/main/java/DataAnalytics.java
@@ -4,7 +4,7 @@ public class DataAnalytics {
 
 	public static void main(String[] args) {
 		
-		Analytics analytics = new Analytics(args[0], args[1], args[2], args[3], args[4], args[5], args[6]);
+		Analytics analytics = new Analytics(args[0], args[1], args[2], args[3], args[4], args[5], args[6], args[7]);
 		analytics.run();
 		
 	}
